app/requirements.txt
selenium
discord.py
icecream
python-dotenv
httpx
chromedriver